refactor: move sqlite store into providers package
